Berkeley DB 12cR1 (12.1.6.1.38)
<https://download.oracle.com/otndocs/products/berkeleydb/html/changelog_6_1.html>

db-6.1.38.tar.gz

+++ src/log/log_print.c
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 1996, 2015 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 1996, 2019 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -154,7 +154,8 @@ __log_print_record(env, recbuf, lsnp, name, spec, info)
 		case LOGREC_LONGARG:
 			LOGCOPY_64(env, &ulltmp, bp);
 			__db_msgadd(env, &msgbuf,  "\t%s: ", sp->name);
-			__db_msgadd(env, &msgbuf,  "%llu", ulltmp);
+			__db_msgadd(env,
+			    &msgbuf,  "%llu", (unsigned long long)ulltmp);
 			__db_msgadd(env, &msgbuf,  "\n");
 			bp += sizeof(ulltmp);
 			break;
